Re-structure how the helpers integrate with the Intl system and app code
This simplifies how the helpers are used to create a composition system
for app code to use.

An `{{#intl}}` block is used to setup a the `data.intl` which will
propogate through all helpers and partials. Once setup, the
`{{intlNumber}}`, `{{intlDate}}`, and `{{intlMessage}}` helpers can be
used and they don't need to be concerned with setting up `locales`.

    {{#intl locales=intl.locale messages=intl.messages}}
        {{intlMessage intlName="USERS_BOOKS"
            firstName=user.firstName lastName=user.lastName
            numBooks=user.numBooks dateBooks=user.dateBooks}}

        {{intlMessage "The number is: {num, number, integer}" num=2000}}

        {{inltNumber 1000 style="currency" currency="USD"}}
    {{/intl}}